third_party/blink/web_tests/http/tests/security/script-onerror-no-crossorigin-cors-expected.txt

CONSOLE ERROR: Uncaught SomeError
The test passes if 'window.onerror' gets sanitized information about an exception thrown in a script loaded without a 'crossorigin' attribute, and delivered with valid CORS headers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS msg is "Script error."
PASS url is ""
PASS line is 0
PASS column is 0
PASS window.errorObject is null
PASS successfullyParsed is true

TEST COMPLETE
